Disabling ca bundle and ca path detection
This time of writing, a single binary built on a specific distribution
is used in all installers of the different distributions. However,
each distribution can deploy their own default ca paths, meaning that
single binary is not portable.

This patch explicitely disables detection.

The plan is: those pe modules that are built over libcurl (http
destination) like stackdriver must explicitely set ca-file and ca-dir
in the http destination driver.

In case of stackdriver, the same detection that is used by libcurl
configure is implemented into the generator.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>